requirements.txt
langgraph
litellm
weaviate-client
sentence-transformers
python-dotenv
langfuse
mem0ai
protobuf==4.23.4
PyMuPDF